Draft of particle effect on multiple elements
